+++ src/app/page.tsx
@@ -1,10 +1,10 @@
-import Hero from '@/components/layout/Hero';
 import { HomeComponent } from '@/components/layout/Home';
+import NewHero from '@/components/layout/NewHero';
 
 export default async function Home() {
   return (
     <main className='flex flex-col items-center justify-between'>
-      <Hero />
+      <NewHero />
       <HomeComponent />
     </main>
   );
