+++ src/components/layout/Home.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { productCardsData } from '@/lib/data/productCardsData';
 import TextSection from './TextSection';
 import { ProductCardsSection } from './ProductCardsSection';
@@ -10,9 +8,7 @@ import {
   newsCardData,
   videosCardData,
 } from '@/lib/data/exampleCardData';
-import { AgentSection } from './AgentSection';
-import { useVerifiedAssistants } from '@/hooks/useAssistants';
-import { Skeleton } from '../ui/skeleton';
+
 import dynamic from 'next/dynamic';
 
 const ExamplesSection = dynamic(() => import('./ExamplesSection'), {
@@ -29,6 +25,9 @@ const PartnersSection = dynamic(() => import('./PartnersSection'), {
 const NumbersSection = dynamic(() => import('./NumbersSection'), {
   ssr: false,
 });
+
+const AgentSection = dynamic(() => import('./AgentSection'));
+
 const headerTextSection = {
   title: 'Chain Agents Live',
   subHeader: 'Explore existing agents or fork one to create your own.',
@@ -42,7 +41,7 @@ const crossSection = {
   thumb: '/examples/agents.webp',
   src: '/video/bitte-sub-natural-language.mp4',
   title: 'Natural Language will Create Most Transactions',
-  subHeader: "It's actaully incredibly easy with smart contracts.",
+  subHeader: "It's actually incredibly easy with smart contracts.",
   factTitle: '',
   fact: '',
   btnTitle: 'Try Now',
@@ -51,35 +50,16 @@ const crossSection = {
 };
 
 export const HomeComponent = () => {
-  const { verifiedAgents: agentData, loading } = useVerifiedAssistants();
   return (
     <>
       <SupportedChainsSection />
       <TextSection {...headerTextSection} />
-      {agentData && agentData?.agents?.length > 0 ? (
-        <AgentSection agentData={agentData} />
-      ) : loading ? (
-        <div className='mb-12'>
-          <div className='flex gap-6 items-center justify-center mb-3'>
-            {Array.from({ length: 5 }).map((_, i) => (
-              <Skeleton key={i} className='w-[305px] h-[75px]' />
-            ))}
-          </div>
-          <div className='flex gap-6 items-center justify-center'>
-            {Array.from({ length: 5 }).map((_, i) => (
-              <Skeleton key={i} className='w-[305px] h-[75px]' />
-            ))}
-          </div>
-        </div>
-      ) : null}
+      <AgentSection />
       <ProductCardsSection data={productCardsData} />
-
       <ExamplesSection data={dropCardData} isVideo={false} />
       <ExamplesSection data={newsCardData} isVideo={false} />
       <ExamplesSection data={videosCardData} isVideo={true} />
-
       <VideoSection {...crossSection} />
-
       <NumbersSection />
       <PartnersSection />
     </>
