+++ src/components/layout/VideoSection.tsx
@@ -1,4 +1,5 @@
 'use client';
+
 import { Button } from '@/components/ui/button';
 
 const VideoSection = ({
